Refactor Action with MVC pattern
